Move importer to components
and move step and after upload logic from uploader into importer
component
