+++ src/app.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 
-import Importer from './importer.jsx'
-import configureStore from './configureStore.js';
+import ImporterLink from './components/importerLink'
+import configureStore from './configureStore';
 
 import {setServerUrl} from './state/server/actions';
 
 const store = configureStore();
 store.dispatch(setServerUrl('http://importer.terranodo.io/'));
-ReactDOM.render(<Provider store={store}><Importer/></Provider>, document.getElementById('main'));
+ReactDOM.render(<Provider store={store}><ImporterLink /></Provider>, document.getElementById('main'));
